Fix couple of reliability problems in the first example
*First*, remove explicit `Exit` messages. They are unnecessary as
channel's closure is a perfectly fine exit message in itself. They are
harmful, because they add a separate code path (as you need to handle
both channel's closure and exit event). They are double harmful,
because these different code-paths correspond to normal and abnormal
termination. Abnormal termination is never tested in practice, and
using a completely separate code path for it is a bad idea, because
that's why how complex distributed systems fail:

https://www.usenix.org/system/files/conference/osdi14/osdi14-paper-yuan.pdf

This also just makes the code shorter, because we can write a `for` loop.

*Second*, make sure that we join worker thread. That makes sure that
don't leak the thread outside of the scope of test, and that we also
propagate it's panic, if any. Note that the current solution with
explicit `.join` is still wrong, because the thread won't be joined on
panic, but to handle that more code/custom crate is required.

*Third*, unwrap all sends. Send only fails if the receiving side has
terminated abormally. In this case, we should escalate failure rather
than continue business as usual, lest we create a deadlock/resource
leak.